Second pass at abstracting webservices and Joomla Integration
