+++ .gitignore
@@ -12,6 +12,9 @@
 .DS_Store
 .idea
 
+#CMS
+staging.joomla.org/
+
 
 # Phing
 build.properties
