+++ src/Api/Soap/Wsdl.php
@@ -10,7 +10,7 @@
 namespace Joomla\Webservices\API\Soap;
 
 use Joomla\Webservices\Uri\Uri;
-use Joomla\Webservices\Api\Hal\HalHelper;
+use Joomla\Webservices\Webservices\ConfigurationHelper;
 use Joomla\Webservices\Xml\XmlHelper;
 
 /**
@@ -273,7 +273,7 @@ class Wsdl
 		$client = XmlHelper::attributeToString($this->webserviceXml, 'client', 'site');
 		$version = !empty($this->webserviceXml->config->version) ? $this->webserviceXml->config->version : '1.0.0';
 		$this->webserviceFullName = $client . '.' . $this->webserviceXml->config->name . '.' . $version;
-		$this->webserviceUrl = HalHelper::buildWebserviceFullUrl($client, $this->webserviceXml->config->name, $version, 'soap');
+		$this->webserviceUrl = ConfigurationHelper::buildWebserviceFullUrl($client, $this->webserviceXml->config->name, $version, 'soap');
 
 		// Root of the document
 		$this->wsdl = new \SimpleXMLElement('<?xml version="1.0" encoding="utf-8" ?><wsdl:definitions'
@@ -316,7 +316,7 @@ class Wsdl
 			// Read list
 			if (isset($this->webserviceXml->operations->read->list))
 			{
-				$filters = HalHelper::getFilterFields($this->webserviceXml->operations->read->list, true, true);
+				$filters = ConfigurationHelper::getFilterFields($this->webserviceXml->operations->read->list, true, true);
 
 				if (empty($filters))
 				{
@@ -361,7 +361,7 @@ class Wsdl
 			{
 				// Add read item messages
 				$inputFields = array_merge(
-					HalHelper::getFieldsArray($this->webserviceXml->operations->read->item, true),
+					ConfigurationHelper::getFieldsArray($this->webserviceXml->operations->read->item, true),
 					array(array('name' => 'language', 'transform' => 'string'))
 				);
 
@@ -382,7 +382,7 @@ class Wsdl
 			if (isset($this->webserviceXml->operations->create))
 			{
 				// Add create messages
-				$inputFields = HalHelper::getFieldsArray($this->webserviceXml->operations->create);
+				$inputFields = ConfigurationHelper::getFieldsArray($this->webserviceXml->operations->create);
 
 				// Add create response messages
 				$outputFields = array(SoapHelper::getResultResource($this->webserviceXml->operations->create));
@@ -394,7 +394,7 @@ class Wsdl
 			if (isset($this->webserviceXml->operations->update))
 			{
 				// Add update messages
-				$inputFields = HalHelper::getFieldsArray($this->webserviceXml->operations->update);
+				$inputFields = ConfigurationHelper::getFieldsArray($this->webserviceXml->operations->update);
 
 				// Add update response messages
 				$outputFields = array(SoapHelper::getResultResource($this->webserviceXml->operations->update));
@@ -406,7 +406,7 @@ class Wsdl
 			if (isset($this->webserviceXml->operations->delete))
 			{
 				// Add delete messages
-				$inputFields = HalHelper::getFieldsArray($this->webserviceXml->operations->delete, true);
+				$inputFields = ConfigurationHelper::getFieldsArray($this->webserviceXml->operations->delete, true);
 
 				// Add delete response messages
 				$outputFields = array(SoapHelper::getResultResource($this->webserviceXml->operations->delete));
@@ -420,7 +420,7 @@ class Wsdl
 				foreach ($this->webserviceXml->operations->task->children() as $taskName => $task)
 				{
 					// Add task messages
-					$inputFields = HalHelper::getFieldsArray($task);
+					$inputFields = ConfigurationHelper::getFieldsArray($task);
 
 					// Add task response messages
 					$outputFields = array(SoapHelper::getResultResource($task));
